refactor DeckLayout component
